smarteye-lam-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
huggingface_hub==0.18.0
loguru==0.7.2

# DocLayout-YOLO dependencies
ultralytics==8.0.196
pydantic==2.4.2